api/chatbot/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
google-generativeai==0.3.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
Django==4.2.0